pyproject.toml
[project]
name = "vod-highlights"
version = "0.1.0"
description = "Local-first Twitch VOD highlight candidate detector"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "faster-whisper>=1.0.0",
  "pyannote.audio>=3.1.0",
  "librosa>=0.10.2",
  "torchaudio>=2.3.0",
  "opencv-python>=4.10.0.84",
  "pandas>=2.2.2",
  "numpy>=1.26.4",
  "scikit-learn>=1.5.1",
  "pyyaml>=6.0.1",
  "pydantic>=2.8.2",
  "typer>=0.12.3",
  "rich>=13.7.1",
  "httpx>=0.27.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "ruff>=0.5.5",
  "mypy>=1.11.1"
]

[project.scripts]
vod-highlights = "src.cli:app"

[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
